+++ packages/electron-builder/CHANGELOG.md
@@ -1,5 +1,22 @@
 # electron-builder
 
+## 26.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`53ee6c6c`](https://github.com/electron-userland/electron-builder/commit/53ee6c6c498a4cc4e64d580c4ec6564137060eae)]:
+  - app-builder-lib@26.0.5
+  - dmg-builder@26.0.5
+
+## 26.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`8b059ad3`](https://github.com/electron-userland/electron-builder/commit/8b059ad3baad440acb0994b2c52f22ea0f1d987f)]:
+  - app-builder-lib@26.0.4
+  - builder-util@26.0.4
+  - dmg-builder@26.0.4
+
 ## 26.0.3
 
 ### Patch Changes